+++ .pre-commit-config.yaml
@@ -50,7 +50,7 @@ repos:
     rev: 6.3.0
     hooks:
       - id: pydocstyle
-        exclude: ^(tests/)|(docs/)|(setup.py)|(pyproject.toml)|(momadm_benchmarks/test/)
+        exclude: ^(tests/)|(docs/)|(setup.py)|(pyproject.toml)|(momaland/test/)
         args:
           - --source
           - --ignore-decorators=override
