--- momadm_benchmarks/envs/__init__.py
@@ -1,2 +1,0 @@
-"""Environment modules for momadm_benchmarks."""
-import momadm_benchmarks.envs.beach_domain
