+++ momaland/envs/__init__.py
@@ -1,0 +1,2 @@
+"""Environment modules for momaland."""
+import momaland.envs.beach_domain
