+++ momaland/envs/beach_domain/mobeach_v0.py
@@ -1,5 +1,5 @@
 """Beach domain environment for multi-objective optimization."""
-from momadm_benchmarks.envs.beach_domain.beach_domain import env, parallel_env, raw_env
+from momaland.envs.beach_domain.beach_domain import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
