+++ momaland/envs/crazyrl/catch/mocatch_v0.py
@@ -1,5 +1,5 @@
 """CrazyRL/Catch environment for MOMARL."""
-from momadm_benchmarks.envs.crazyrl.catch.catch import env, parallel_env, raw_env
+from momaland.envs.crazyrl.catch.catch import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
