+++ momaland/envs/crazyrl/escort/moescort_v0.py
@@ -1,5 +1,5 @@
 """CrazyRL/Escort environment for MOMARL."""
-from momadm_benchmarks.envs.crazyrl.escort.escort import env, parallel_env, raw_env
+from momaland.envs.crazyrl.escort.escort import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
