+++ momaland/envs/crazyrl/surround/mosurround_v0.py
@@ -1,5 +1,5 @@
 """CrazyRL/Surround environment for MOMARL."""
-from momadm_benchmarks.envs.crazyrl.surround.surround import env, parallel_env, raw_env
+from momaland.envs.crazyrl.surround.surround import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
