+++ momaland/envs/item_gathering/asset_utils.py
