+++ momaland/envs/item_gathering/assets/agent.png
