+++ momaland/envs/item_gathering/assets/item.png
