+++ momaland/envs/item_gathering/assets/map_bg1.png
