+++ momaland/envs/item_gathering/assets/map_bg2.png
