+++ momaland/envs/item_gathering/moitemgathering_v0.py
@@ -1,9 +1,5 @@
 """Item Gathering environment for multi-objective optimization."""
-from momadm_benchmarks.envs.item_gathering.item_gathering import (
-    env,
-    parallel_env,
-    raw_env,
-)
+from momaland.envs.item_gathering.item_gathering import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
