+++ momaland/envs/multiwalker/momultiwalker_v0.py
@@ -1,5 +1,5 @@
 """Multiwalker domain environment for MOMARL."""
-from momadm_benchmarks.envs.multiwalker.multiwalker import env, parallel_env, raw_env
+from momaland.envs.multiwalker.multiwalker import env, parallel_env, raw_env
 
 
 __all__ = ["env", "parallel_env", "raw_env"]
