+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools >= 61.0.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "momadm-benchmarks"
+name = "momaland"
 description="A standard API for Multi-Objective Multi-Agent Decision making and a diverse set of reference environments."
 readme = "README.md"
 requires-python = ">= 3.8"
@@ -46,16 +46,16 @@ all = [
 testing = ["pytest ==7.1.3"]
 
 [project.urls]
-Homepage = "https://github.com/rradules/momadm-bechmarks/tree/main"
-Repository = "https://github.com/rradules/momadm-bechmarks/tree/main"
-Documentation = "https://github.com/rradules/momadm-bechmarks/tree/main"
-"Bug Report" = "https://github.com/rradules/momadm-bechmarks/tree/main/issues"
+Homepage = "https://github.com/rradules/momaland/tree/main"
+Repository = "https://github.com/rradules/momaland/tree/main"
+Documentation = "https://github.com/rradules/momaland/tree/main"
+"Bug Report" = "https://github.com/rradules/momaland/tree/main/issues"
 
 [tool.setuptools]
 include-package-data = true
 
 [tool.setuptools.packages.find]
-include = ["momadm_benchmarks", "momadm_benchmarks.*"]
+include = ["momaland", "momaland.*"]
 
 [tool.setuptools.package-data]
 mo_gymnasium = [
@@ -74,14 +74,14 @@ include = '\.pyi?$'
 [tool.isort]
 atomic = true
 profile = "black"
-src_paths = ["momadm_benchmarks", "tests", "docs/scripts"]
+src_paths = ["momaland", "tests", "docs/scripts"]
 extra_standard_library = ["typing_extensions"]
 indent = 4
 lines_after_imports = 2
 multi_line_output = 3
 
 [tool.pyright]
-include = ["momadm_benchmarks/**", "tests/**"]
+include = ["momaland/**", "tests/**"]
 exclude = ["**/node_modules", "**/__pycache__"]
 strict = []
 
@@ -115,6 +115,6 @@ reportUnboundVariable = "warning"
 [tool.pytest.ini_options]
 filterwarnings = ['ignore:.*The environment .* is out of date.*']
 norecursedirs = [
-    "momadm_benchmarks",
+    "momaland",
 ]
 # filterwarnings = ['ignore:.*step API.*:DeprecationWarning']
